+++ packages/core/src/middleware/global-error-handler.spec.ts
@@ -6,7 +6,6 @@ describe('Middleware / GlobalErrorHandlerMiddleware', () => {
   let middleware: Partial<MiddlewareDescriptor>
   const params: MiddlewareParams = {
     clientId: 'testClient',
-    context: {},
     resourceMethod: 'bar',
     resourceName: 'Foo',
   }